portfolio/permissions.py
from rest_framework.permissions import BasePermission, SAFE_METHODS


class IsAdminOrOwner(BasePermission):
    def has_object_permission(self, request, view, obj):
        return bool(
            request.method in SAFE_METHODS and
            request.user and (
                request.user.is_superuser or request.user == obj.user
            ) 
        )