pyproject.toml
[tool.poetry]
name = "lanars"
version = "0.1.0"
description = ""
authors = ["Yaroslav Martynenko <stikblacklabel@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.2.0"}
drf-yasg = "^1.21.3"
gunicorn = "^20.1.0"
django-environ = "^0.9.0"
Pillow = "^9.2.0"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
